fix: refactor mem-resources-boot to resolve eslint line limit
