+++ src/resources/embedded-mcp-resources.ts
@@ -18,7 +18,7 @@ export const mcpResources = {
     "activate": "Rank stored **adapters** for the user’s intent and return the literal next\n`forward` call arguments.\n\n**When to call:** Whenever the message describes an action, task, or workflow —\neven if the user never says \"KAIROS\". Always pass a short `query` summary.\n\nHappy-path flow:\n\n1. Call `activate` with a short `query`.\n2. Pick one `choices[]` row.\n3. Copy `choices[].forward_first_call` directly into `forward`.\n\nExample:\n\n```json\n{\n  \"query\": \"review implementation plan\"\n}\n```\n\nResponse row (match/refine):\n\n```json\n{\n  \"role\": \"match\",\n  \"uri\": \"kairos://adapter/phase-critic\",\n  \"forward_first_call\": {\n    \"uri\": \"kairos://adapter/phase-critic\"\n  }\n}\n```\n\nNext call:\n\n```json\n{ \"uri\": \"kairos://adapter/phase-critic\" }\n```\n\n**Input**\n\n- `query` — required short intent summary (about 3-8 words).\n- `execution_id` (optional) — server-issued id for a refinement sequence. Omit on the first activate call; echo it back when you re-run activate with a refined query for the same user request.\n- `space` / `space_id` (optional) — narrow search to one space: `\"personal\"`, a\n  full group path such as `\"/shared/pe-team\"` (optional `\"Group: \"`\n  prefix), or your raw `space_id` (same forms as **`train`** / **`tune`** `space`).\n- `max_choices` (optional) — cap on match rows returned.\n\n**Output:** Always `must_obey: true`. Includes `choices` (each with slug-form\n`uri`, `label`, `adapter_name`, `activation_score`, `role`, `tags`,\n`next_action`, optional `adapter_version`, optional `activation_patterns`,\n`space_name`, `slug`, and `forward_first_call`) plus `execution_id` (echo it back when refining).\n\n`forward_first_call` is required on `match` and `refine`, and `null` on\n`create`.\n\n**`kairos_local_artifact_dir`** — ordered URI hints (preferred first) for the run's local handoff dir (drafts, review outputs, generated files, checksums shared between layers and subagents). Two schemes, both resolved **on your machine**: `project://<rel>` → `<your project root>/<rel>`; `user://<rel>` → `<your home or $XDG_CONFIG_HOME>/<rel>`. Pick `project://` when you have exactly one project context; fall through to `user://` when your session spans multiple projects so artifacts from different projects don't collide. After resolving, `export KAIROS_LOCAL_ARTIFACT_DIR=\"<absolute>\"` for shell challenges (skip if your shell already defines it). The server never resolves these to a path on its own filesystem; the value carries no server paths and is identical for stdio and HTTP transports. Not a shell cwd, not a Docker `WORKDIR`.\n\nWhen several spaces contain similar adapters, the server prefers your **default write space** (usually **Personal**) on ties so a personal copy can override a group template.\n\n**Roles**\n\n- `match` — `activation_score` is a normalized 0.0–1.0 confidence score; copy\n  `forward_first_call` into `forward`.\n- `refine` — guided help to improve the query; copy `forward_first_call` into\n  `forward`.\n- `create` — no stored adapter matched; **`train`** new adapter markdown (adapter/workflow creation flow).\n\n**Rules**\n\n- Pick **one** choice and obey **that** choice’s `next_action` (not a different URI).\n- Weak matches (e.g. all scores &lt; 0.5): prefer the refine choice once before creating.\n- If you re-run `activate` more than twice with the same `execution_id`, the default refine footer is removed so you stop looping and instead create a custom adapter/workflow or ask the user clarifying questions.\n",
     "delete": "Remove adapters or individual layers by URI.\n\n**Input**\n\n- `uris` — non-empty array of `kairos://adapter/{uuid}` or\n  `kairos://layer/{uuid}`.\n\n**Behavior**\n\n- **Adapter URI** — deletes **all** layers belonging to that adapter.\n- **Layer URI** — deletes that layer memory only.\n\n**Output:** Per-URI `status` (`deleted` | `error`), `message`, and aggregate counts.\n\n**Warning:** Deletion is destructive. Prefer **`export`** before removing production adapters.\n",
     "export": "Export an **adapter**, a single **layer**, an **artifact** (`source`), or a **skill-shaped bundle** for backup, inspection, training pipelines, or portable skill distribution.\n\n**Input**\n\n- **Selection** — exactly one of:\n  - **`uri`** — `kairos://adapter/{uuid|slug}`, `kairos://layer/{uuid}` (optional `execution_id` on layer URIs where applicable), or `kairos://artifact/{uuid|slug}` for source/artifact payloads.\n  - **`adapters`** — non-empty array of adapter identifiers (full URIs and/or bare slugs). Produces one bundle with one top-level folder per exported skill **slug** (capped per server policy).\n  - **`all_adapters: true`** with **`space_name`** — export every adapter in that space (human-readable space **name**, not a raw space id).\n- **`format`** (optional) — default **`skill_zip`** (skill-shaped ZIP delivered through **`download_ref`** by default). Other values:\n  - **`markdown`** — flat serialized adapter Markdown (single **`uri`** only; smallest payload for **`tune`** or a quick local file). Prefer **`skill_tree`** / **`skill_zip`** for portable **`SKILL.md` + `artifacts/` + integrity.\n  - **`skill_tree`** — JSON tree of generated files (paths and UTF-8 contents).\n  - **`trace_jsonl`**, **`reward_jsonl`**, **`sft_jsonl`**, **`preference_jsonl`** — training datasets (single **`uri`** only).\n  - **`source`** — artifact listing or payload (single **`uri`**).\n- **`include_reward`** (optional, default **`true`**) — include reward fields when serializing **`trace_jsonl`**. **`reward_jsonl`** always serializes reward data and skips unrewarded executions.\n- **`delivery`** (optional, **`skill_zip`** only) — default **`download_ref`** returns compact JSON with a short-lived download URL. Use **`inline_base64`** only when the ZIP must be embedded in the JSON response.\n\n**Output:** `content` (string), `content_type`, `format`, optional `item_count`, adapter metadata. Default **`skill_zip`** responses set **`download_ref`** with `url`, `expires_at`, `filename`, and `content_type`; fetch the URL with **`curl -fLOJ`** or equivalent to get ZIP bytes. **`inline_base64`** responses may set **`content_encoding: \"base64\"`**, **`bundle_sha256`**, **`skill_bundle_manifest`** (JSON string), and **`export_adapter_count`**. Server operators may set **`KAIROS_EXPORT_ZIP_COMPRESSION_LEVEL`** (**0**–**9**, default **6**) to tune **`skill_zip`** CPU vs size (**0** = store-only; **9** = slowest, smallest).\n\n**HTTP bytes:** **`POST /api/export`** returns JSON for **`skill_zip`**. ZIP bytes are served only by **`GET /export/skill-zip/{opaque}`** from **`download_ref.url`**.\n\n**Skill-shaped export:** Serialized tree uses **`contract`** in JSON fences (not `challenge`), **Activation Patterns** / **Reward Signal** headings where applicable, and optional **`<slug>/artifacts/`** files from storage. Each **`<slug>/`** folder also contains **`SHA256SUMS`** (GNU **`sha256sum`** text, paths relative to that folder) listing **`SKILL.md`** and every **`artifacts/`** file. **`skill_tree`** includes a per-skill **`diagnostics`** array (for example extension vs MIME warnings on attached files, or **`artifact_stored_sha_mismatch`** when a stored artifact digest disagrees with exported bytes).\n\n**Training formats:** **`reward_jsonl`** emits only rows with stored reward data. **`sft_jsonl`** and **`preference_jsonl`** apply their export gates. Ungraded rewards stay in **`trace_jsonl`** but are excluded from model-training formats where documented.\n\n**RFT gate:** `rft_jsonl` is intentionally not exposed yet.\n\n**Use with `tune`:** Edit **flat** exported Markdown, then **`tune`** with matching `uris` / `content`. Obtain flat Markdown via **`format: markdown`** (single adapter or layer **`uri`**).\n\n**Local install (agents):** **`export`** is also used to **mirror** stored content for execution (scripts under **`artifacts/`**, **`SHA256SUMS`** checks). For a **full** tree, use **`skill_zip`** and unzip; a **canonical** install root on the same machine as the CLI config is **`$XDG_CONFIG_HOME/kairos/skills/<slug>/`** (Unix; see `getKairosSkillInstallDirForSlug` in server sources / [`workflow-export.md`](../../../docs/architecture/workflow-export.md)). For **only** the adapter document, **`format: markdown`** and write **`content`** to your chosen path under that tree or the workspace.\n",
-    "forward": "Run an **adapter** layer by layer.\n\nFirst call (start a run):\n\n```json\n{ \"uri\": \"kairos://adapter/<slug>\" }\n```\n\nCopy this from `activate.choices[].forward_first_call.uri`. Adapter URIs are\nslug-only on the wire.\n\nContinuation call:\n\n```json\n{\n  \"uri\": \"kairos://layer/<uuid>?execution_id=<uuid>\",\n  \"solution\": {\n    \"type\": \"<contract.type>\",\n    \"outcome\": \"success\",\n    \"evidence\": { /* type-specific proof data */ },\n    \"nonce\": \"<contract.nonce when present>\",\n    \"proof_hash\": \"<contract.proof_hash when present>\"\n  }\n}\n```\n\nCopy `forward.next_call.args` from the previous response, then fill the\n`solution.evidence` payload. Layer URIs are UUID-only; never substitute a slug\nfor a layer URI.\n\nOutput always includes:\n\n- `contract` for the current layer\n- `next_action` prose guidance\n- `next_call` (authoritative machine-readable next call)\n\n`next_call.kind` is:\n\n- `forward` on non-terminal layers\n- `reward` on terminal layers\n\n## Unified solution envelope (v2 format)\n\nEvery solution, regardless of type, follows the same envelope:\n\n```json\n{\n  \"type\": \"<challenge_type>\",\n  \"outcome\": \"success | failure | skipped\",\n  \"evidence\": { \"/* type-specific proof */\": true },\n  \"nonce\": \"<echo from contract>\",\n  \"proof_hash\": \"<echo from contract>\"\n}\n```\n\n### Per-type `evidence` shapes\n\n#### shell\n\n```json\n{\n  \"type\": \"shell\",\n  \"outcome\": \"success\",\n  \"evidence\": {\n    \"exit_code\": 0,\n    \"stdout\": \"...\",\n    \"stderr\": \"\",\n    \"duration_seconds\": 1.2\n  }\n}\n```\n\n#### mcp\n\n```json\n{\n  \"type\": \"mcp\",\n  \"outcome\": \"success\",\n  \"evidence\": {\n    \"tool_name\": \"editJiraIssue\",\n    \"arguments\": { \"issueIdOrKey\": \"BIB-1383\" },\n    \"response\": { \"key\": \"BIB-1383\", \"webUrl\": \"...\" }\n  }\n}\n```\n\n#### comment\n\n```json\n{\n  \"type\": \"comment\",\n  \"outcome\": \"success\",\n  \"evidence\": {\n    \"text\": \"Verified all 78 issues now have repo: labels.\"\n  }\n}\n```\n\n#### user_input\n\n```json\n{\n  \"type\": \"user_input\",\n  \"outcome\": \"success\",\n  \"evidence\": {\n    \"confirmation\": \"agreed. proceed\",\n    \"timestamp\": \"2026-05-18T14:49:00Z\"\n  }\n}\n```\n\n#### tensor\n\n```json\n{\n  \"type\": \"tensor\",\n  \"outcome\": \"success\",\n  \"evidence\": {\n    \"name\": \"priority_score\",\n    \"value\": [0.85, 0.72, 0.91]\n  }\n}\n```\n\n## Older format acceptance\n\nThe server accepts both old (v1) and new (v2) formats silently:\n\n### Old (v1) — accepted forever, never advertised\n\n```json\n{\n  \"type\": \"mcp\",\n  \"mcp\": { \"tool_name\": \"x\", \"result\": {}, \"success\": true }\n}\n```\n\n### New (v2) — advertised in error messages, docs, examples\n\n```json\n{\n  \"type\": \"mcp\",\n  \"outcome\": \"success\",\n  \"evidence\": { \"tool_name\": \"x\", \"response\": {} }\n}\n```\n",
+    "forward": "Run an **adapter** layer by layer.\n\nFirst call (start a run):\n\n```json\n{ \"uri\": \"kairos://adapter/<slug>\" }\n```\n\nCopy this from `activate.choices[].forward_first_call.uri`. Adapter URIs are\nslug-only on the wire.\n\nContinuation call:\n\n```json\n{\n  \"uri\": \"kairos://layer/<uuid>?execution_id=<uuid>\",\n  \"solution\": {\n    \"type\": \"<contract.type>\",\n    \"outcome\": \"success\",\n    \"evidence\": { /* type-specific proof data */ },\n    \"nonce\": \"<contract.nonce when present>\",\n    \"proof_hash\": \"<contract.proof_hash when present>\"\n  }\n}\n```\n\nCopy `forward.next_call.args` from the previous response, then fill the\n`solution.evidence` payload. Layer URIs are UUID-only; never substitute a slug\nfor a layer URI.\n\nOutput always includes:\n\n- `contract` for the current layer\n- `next_action` prose guidance\n- `next_call` (authoritative machine-readable next call)\n\n`next_call.kind` is:\n\n- `forward` on non-terminal layers\n- `reward` on terminal layers\n\n## Unified solution envelope (v2 format)\n\nEvery solution, regardless of type, follows the same envelope:\n\n```json\n{\n  \"type\": \"<challenge_type>\",\n  \"outcome\": \"success | failure | skipped\",\n  \"evidence\": { \"/* type-specific proof */\": true },\n  \"nonce\": \"<echo from contract>\",\n  \"proof_hash\": \"<echo from contract>\"\n}\n```\n\n### Per-type `evidence` shapes\n\n#### shell\n\n```json\n{\n  \"type\": \"shell\",\n  \"outcome\": \"success\",\n  \"evidence\": {\n    \"exit_code\": 0,\n    \"stdout\": \"...\",\n    \"stderr\": \"\",\n    \"duration_seconds\": 1.2\n  }\n}\n```\n\n#### mcp\n\n```json\n{\n  \"type\": \"mcp\",\n  \"outcome\": \"success\",\n  \"evidence\": {\n    \"tool_name\": \"editJiraIssue\",\n    \"arguments\": { \"issueIdOrKey\": \"BIB-1383\" },\n    \"response\": { \"key\": \"BIB-1383\", \"webUrl\": \"...\" }\n  }\n}\n```\n\n#### comment\n\n```json\n{\n  \"type\": \"comment\",\n  \"outcome\": \"success\",\n  \"evidence\": {\n    \"text\": \"Verified all 78 issues now have repo: labels.\"\n  }\n}\n```\n\n#### user_input\n\n```json\n{\n  \"type\": \"user_input\",\n  \"outcome\": \"success\",\n  \"evidence\": {\n    \"confirmation\": \"agreed. proceed\",\n    \"timestamp\": \"2026-05-18T14:49:00Z\"\n  }\n}\n```\n\n#### tensor\n\n```json\n{\n  \"type\": \"tensor\",\n  \"outcome\": \"success\",\n  \"evidence\": {\n    \"name\": \"priority_score\",\n    \"value\": [0.85, 0.72, 0.91]\n  }\n}\n```\n\n## Format migration and compatibility\n\nThe server accepts both old (v1) and new (v2) formats silently:\n\n### Old (v1) — accepted forever, never advertised\n\n```json\n{\n  \"type\": \"mcp\",\n  \"mcp\": { \"tool_name\": \"x\", \"result\": {}, \"success\": true }\n}\n```\n\n### New (v2) — advertised in error messages, docs, examples\n\n```json\n{\n  \"type\": \"mcp\",\n  \"outcome\": \"success\",\n  \"evidence\": { \"tool_name\": \"x\", \"response\": {} }\n}\n```\n\n### When to use v2 format\n\n**Always prefer v2** for new protocols and solutions. Use v2 when:\n- Creating new adapter protocols\n- Submitting solutions to any challenge type\n- Reading error messages (they show v2 examples)\n- Following `next_action` guidance (it teaches v2 format)\n\n### Key differences between v1 and v2\n\n| Aspect | v1 (older) | v2 (unified) |\n|--------|-------------|--------------|\n| **Structure** | `{type, [type]: {...}}` | `{type, outcome, evidence: {...}}` |\n| **Success signal** | `success: true` inside type object | `outcome: \"success\"` at top level |\n| **MCP response** | `result` field | `response` field inside evidence |\n| **Consistency** | Different shapes per type | Unified envelope for all types |\n\n### Automatic format conversion\n\nThe server automatically converts v1 to v2 internally:\n- v1 `mcp.result` becomes v2 `evidence.response`\n- v1 `mcp.success` is derived from v2 `outcome` if not provided\n- Older type-specific objects are wrapped in `evidence` envelope\n\n### Mixed format protocols\n\nYou can safely mix v1 and v2 solutions within the same protocol run. The server handles both formats seamlessly:\n\n```json\n// Step 1: v1 format (still accepted)\n{\"type\": \"shell\", \"shell\": {\"exit_code\": 0, \"stdout\": \"done\"}}\n\n// Step 2: v2 format (preferred)\n{\"type\": \"mcp\", \"outcome\": \"success\", \"evidence\": {\"tool_name\": \"search\", \"response\": {...}}}\n\n// Step 3: v1 format (still works)\n{\"type\": \"comment\", \"comment\": {\"text\": \"verification complete\"}}\n```\n\nAll examples in this documentation use v2 format. Follow these examples for best results.\n",
     "reward": "Finalize an adapter run with an outcome and optional evaluator metadata.\n\n**When:** After **`forward`** reports the last layer is done and `next_action` tells you to call **`reward`**.\n\n**Input**\n\n- `uri` — **`kairos://layer/{uuid}`** (final layer; include `?execution_id=...` when the run used it).\n- `outcome` — `success` or `failure`.\n- Optional: `score` (0–1), `feedback`, `rater`, `rubric_version`, `llm_model_id`.\n\n**Output:** Per-layer result rows with `rated_at`, normalized grading\nmetadata (`grader_kind`, `evaluation_label`), export eligibility\nflags plus blocker lists for `sft_jsonl` and `preference_jsonl`, and\n`next_call: null` as the explicit terminal marker.\n\n**Rules:** Use only layer URIs returned by **`forward`**. Do not substitute adapter URIs here unless the tool schema explicitly allows it (this tool expects a layer URI).\n",
     "spaces": "List **spaces** available to the caller (human-readable names) and how many\nadapters each contains.\n\n**Input**\n\n- `include_adapter_titles` (optional) — when true, include per-space adapter\n  titles and layer counts.\n- `include_widget_html` (optional) — when true, load adapter titles and append a\n  second tool content part with an HTML table (type badges, expandable adapter\n  lists) for hosts that render HTML.\n\n**Output:** `spaces` array with `name`, `space_id`, `type` (`personal` | `group` | `app` | `other`), `adapter_count`, and optionally `adapters` (`adapter_id`, `title`, `layer_count`).\n\nWhen running in simple mode (`AUTH_ENABLED=false`), `spaces` includes both a writable **Personal** space and the read-only **Kairos app** space.\n\n**Collaboration patterns**\n\n1. **Precedence** — `activate` search ties break in favour of your **default write space** (usually **Personal**). A personal fork can rank beside an identical-scoring group adapter.\n2. **Scoping `activate`** — pass `space` / `space_id` using the same strings as **`train`** / **`tune`**: `\"personal\"`, a full group path such as `\"{{KAIROS_GROUP_SPACE_PATH_EXAMPLE}}\"` (optional `\"Group: \"` prefix), or the raw `space_id` from this tool.\n3. **Fork** — use **`train`** with `source_adapter_uri` (`kairos://adapter/{uuid}`) and target `space` to copy markdown into a **new** adapter UUID in another space; the original is unchanged.\n4. **Move** — use **`tune`** with `space` (and optional content edits) to reassign an existing adapter’s layers to another allowed space.\n5. **Visibility** — `activate` **match** choices include **`space_name`** so you can see whether a hit is personal, group, or app-level.\n\n**KAIROS tool order** for runs: **`activate`** → **`forward`** (per layer until `next_action` → **`reward`**) → **`reward`**. Use this tool first when you need valid **`space`** values or an inventory before activation.\n",
     "train": "Register a new **adapter** from markdown (one H1 = one adapter; each verifiable segment ends with a fenced `json` block containing a top-level **`contract`** object).\n\n**Input**\n\n- `content` — full document string (optional if `source_adapter_uri` supplies content).\n- `llm_model_id` — required model identifier string.\n- `force_update` (optional) — replace an existing adapter with the same label.\n- `protocol_version` (optional) — version string (e.g. semver) stored on the adapter.\n- `space` (optional) — `\"personal\"` or a full group path such as `\"{{KAIROS_GROUP_SPACE_PATH_EXAMPLE}}\"` (target space for the new adapter).\n- `source_adapter_uri` (optional) — `kairos://adapter/{uuid}` to **fork**: export that adapter’s markdown and **train** a **new** adapter (new ids). If you also pass `content`, that text is used instead of the export (customize before calling **`train`**).\n- `mime` (optional) — content MIME type. Defaults to `text/markdown`.\n- `artifact_name` (required when `mime` is not `text/markdown`) — file name shown when listing adapter artifacts.\n- `adapter_uri` (required when `mime` is not `text/markdown`) — `kairos://adapter/{uuid}` or `kairos://adapter/{slug}` target adapter for the artifact attachment.\n- `relative_path` (optional, non-markdown artifacts only) — path relative to the skill bundle root (forward slashes; no `..` segments). When set, **`export`** **`skill_zip`** and **`skill_tree`** place the file at this path instead of the default **`artifacts/<sanitized artifact name>`** layout.\n\n**Required structure (validated before store)**\n\n- H1 title for the adapter.\n- First H2: **Activation Patterns**.\n- Last H2: **Reward Signal**.\n- At least one fenced **` ```json `** block whose JSON has a top-level\n  **`contract`** object.\n- Contract fences must use the `json` language tag only (no plain ``` blocks holding contract JSON).\n- Non-markdown artifacts (`mime` not `text/markdown`) are stored as adapter-linked artifacts instead of stored Markdown layers.\n\n**Size limits (safety):** Before structure checks, the server rejects oversized Markdown and artifact bodies (max logical lines, max UTF-8 bytes per line, and a total-byte ceiling from env — default line cap **350**, default per-line bytes **8192**, safety factor **1.15**). See architecture [Train workflow](../../../docs/architecture/workflow-train.md#adapter-markdown-size-limits-safety).\n\n**Output:** `status: stored` and `items` with `layer_uuid`, `adapter_uri`, `layer` URIs, labels, tags.\n\n**After train:** Use **`activate`** / **`forward`** to execute; use **`tune`**\nfor in-place adapter/layer edits (non-structural); use **`train`** with\n`force_update: true` for structural adapter replacement; use **`export`** to\ndump markdown or datasets.\n\nFor contract shapes and examples, call **`forward`** with\n`kairos://adapter/create-new-protocol` and omit `solution` on the first call of\nthe run (then follow `next_action`).\n",
@@ -29,7 +29,7 @@ export const mcpResources = {
     "create-new-protocol": "---\nslug: create-new-protocol\nversion: \"4.5.1\"\ntitle: Create / Review / Refactor KAIROS Protocol\n---\n\n\n\n# Create / Review / Refactor KAIROS Protocol\n\n> **Why we create protocols — a note to the agent reading this.**\n>\n> KAIROS protocols are not documentation. They are not conventional apps.\n> They are **interfaces for a fundamentally new kind of user: AI agents.**\n>\n> You, the agent reading this, are the primary user. When you execute a\n> protocol, you follow instructions designed to help you harmonise with the\n> human you serve. Every phase boundary, every contract, every explicit stop\n> condition exists because you cannot reliably infer these on your own — and\n> honest uncertainty is always better than confident drift.\n>\n> When you author a new protocol, you are designing behaviour for yourself\n> and for other agents — some weaker, some stronger, all trying to serve\n> their humans well. Write for them:\n>\n> - **Single source of truth.** Never duplicate rules across protocols.\n>   When rules drift between files, agents execute contradictory behaviour\n>   depending on entry point — and they cannot detect the contradiction.\n> - **Explicit over implicit.** Agents fill ambiguity with hallucination.\n>   Say exactly what to do, when to stop, and when to ask.\n> - **Phases teach modes.** Each step teaches the agent which cognitive mode\n>   it is in: explore, clarify, decide, execute, verify, review. Blurring\n>   these modes causes the agent to act when it should verify, or guess\n>   when it should ask.\n> - **Truth over fluency.** A protocol that reads well but causes wrong\n>   behaviour is worse than one that reads rough but executes correctly.\n\n## Activation Patterns\n\nKAIROS authoring lifecycle — create, review, or fix adapters / protocols / workflows.\nIn this context, those words refer to the same stored artifact and can be treated\nas synonyms. Supports three operations:\n\n- **create** — draft and train a new adapter / protocol from scratch.\n- **review** — audit an existing adapter, family, or bundle for structural issues,\n  missing cross-references, chaining gaps, DRY violations, stale guidance, or\n  agent-hostile ambiguity. Read-only.\n- **fix** — apply specific corrections to an existing adapter / protocol and re-train.\n\n**Run this protocol when the user says ANY of:**\n\n- \"create a new KAIROS protocol\" / \"create new protocol adapter\"\n- \"register personal KAIROS adapter\" / \"register a personal adapter\"\n- \"train adapter into personal\" / \"train protocol into personal\"\n- \"train a workflow\" / \"register a new adapter\" / \"create a new workflow\"\n- \"review protocol\" / \"audit adapters\" / \"check protocol families for gaps\"\n- \"fix protocol\" / \"update adapter\" / \"patch protocol\" / \"refactor protocol\"\n- \"create new protocol\" (when `activate` found no match and user confirms)\n\n**Trigger pattern:** **create** / **train** / **register** / **review** /\n**audit** / **fix** / **update** / **patch** / **refactor** +\n(protocol / adapter / workflow).\n\n**Must Never:**\n- Run when the user only asked to execute an existing protocol.\n- Mutate adapters during a `review` operation.\n- Optimise only for pretty wording; this protocol designs agent behaviour.\n\n**Must Always:**\n- Infer `{operation}` from trigger: create / review / fix.\n- Optimise for truth, low ambiguity, explicit sequencing, and user-need alignment.\n- Design or review phase transitions explicitly: discovery, clarification,\n  ideation, planning, execution, validation, review, handoff.\n- Include Activation Patterns as the first H2 and Reward Signal as the last H2\n  in any newly drafted adapter.\n- Enforce the 350-line limit per file.\n\n**Good trigger examples:**\n- \"Create a new KAIROS protocol for code review\" → run this protocol (create)\n- \"No match found; I want to create a new protocol\" → run this protocol (create)\n- \"Review all v4 protocol families for missing cross-references\" → run this protocol (review)\n- \"Refactor the MR adapter to be less ambiguous\" → run this protocol (fix)\n\n**Bad trigger examples:**\n- \"Run the standardise project protocol\" → use activate / forward\n- \"Search for deployment protocol\" → use activate\n- \"Create an MR\" → use the MR adapter, not this protocol\n\n## Preflight Dependencies\n\nBefore running authoring actions, verify dependency and auth readiness for this\nrun:\n\n1. Required MCP tools are available and reachable: `activate`, `forward`,\n   `reward`, and the operation-dependent tool (`train`, `export`, or `tune`).\n2. Required CLI tools for planned shell steps are installed and callable.\n3. Required auth/session state is valid for the target MCP server.\n\nIf any check fails, stop before main execution and ask the user to choose one:\n- proceed with AI-assisted setup and then resume\n- abort this protocol run now\n\nDo not continue to intent confirmation until preflight passes.\n\n```json\n{\"contract\":{\"type\":\"comment\",\"comment\":{\"min_length\":80},\"required\":true}}\n```\n\n## Confirm Intent\n\nInfer `{operation}` from the user's trigger:\n\n| Signal | Operation |\n|---|---|\n| \"create\", \"new\", \"train\", no match from activate | **create** |\n| \"review\", \"audit\", \"check\", \"gaps\", \"cross-references\" | **review** |\n| \"fix\", \"update\", \"patch\", \"refactor\", \"add missing\", \"correct\" | **fix** |\n\nFor `create`: ask whether to create a new protocol or refine the search.\nFor `review`: ask which adapter(s), family, or bundle to review.\nFor `fix`: ask which adapter to fix and what the issue is.\n\n```json\n{\"contract\":{\"type\":\"user_input\",\"user_input\":{\"prompt\":\"Confirm operation and target\"},\"required\":true}}\n```\n\n## Gather Requirements\n\n### For `create` — collect details for a new protocol\n\n**1. Determine the pattern:**\n\n| Signal | Pattern |\n|---|---|\n| One workflow, no variants | **Standalone** — one protocol, all steps inline |\n| Multiple issue types / output formats / media | **Router + Extensions** — router classifies, routes via `activate` |\n| Domain has a decision tree | **Router + Extensions** |\n| Single workflow but exceeds 350 lines or too complex for weaker models | **Chain** — split into sequential adapters linked via `forward` + slug |\n| Reference material needed by steps | Extract to a separate file, not a protocol |\n\n**2. Collect per protocol:**\n\n- **Title** — clear, descriptive H1.\n- **User problem** — what real problem does this solve?\n- **Agent behaviour** — what behaviour should this create in the agent?\n- **Mode map** — which steps are for discovery, clarification, ideation,\n  planning, execution, validation, review, or handoff?\n- **Steps** — each H2 must earn its place and end with a challenge.\n- **Challenge type per step** — use the strongest observable contract possible.\n- **Domain reference** — guides / docs the protocol should load.\n- **Existing protocols** — reuse, compose, or extract shared content.\n- **DRY check** — extract duplicated content.\n- **Size check** — if any file may exceed 350 lines, split now.\n- **Chain position** — is this a chain entry, chain middle, or chain exit?\n  If chain middle/exit, what slug does the previous link forward to?\n- **Failure analysis** — where would a weaker agent misread this protocol?\n\n**3. If Router pattern:**\n- What are the variants?\n- What decision questions classify the variant?\n- What types are forbidden?\n- Which parts stay in the router, and which belong only in extensions?\n\n**4. If Chain pattern:**\n\nLoad `protocol-linking-guide` for full chain structuring rules\n(`forward` with `kairos://adapter/protocol-linking-guide`).\n\nCollect:\n- What are the chain links and their concerns?\n- What is the minimum number of links that helps weaker models?\n- Which link is the chain entry point (matched by `activate`)?\n- What artifacts flow between links (context set in earlier phases)?\n\n### For `review` — audit existing adapters against this checklist\n\nExport target adapter(s). For each adapter, check:\n\n1. **User-need fit** — does the adapter solve a real user problem, or merely\n   mirror literal wording?\n2. **Mode / phase clarity** — are discovery, clarification, ideation,\n   execution, validation, and review clearly separated?\n3. **Cross-reference completeness** — are all required related adapters linked?\n4. **Adapter handoff gaps** — does it hand off where another adapter should own\n   the work?\n5. **DRY violations** — is policy content duplicated instead of linked?\n6. **Contract coverage** — does every executable H2 have the right contract?\n7. **Truth-first behaviour** — does it reduce guessing, fake certainty, and\n   silent assumption-making?\n8. **Activation pattern accuracy** — do triggers match what users actually say?\n9. **Structure compliance** — first H2 Activation Patterns, last H2 Reward\n   Signal, no duplicate H2s, under 350 lines.\n10. **Sibling consistency** — are equivalent steps present across sibling\n    adapters in the same family?\n11. **Chained adapter families** (when applicable) — extension adapters include\n    `chain_root` and a prerequisite clause pointing at the root; parent adapters\n    use explicit `forward` when the next adapter slug is already known; reserve\n    `activate` for ambiguous runtime discovery (see `protocol-linking-guide`,\n    **Chaining Back and Forward**).\n\n### For `fix` — identify the specific issue\n\nExport the target adapter, read the current markdown, identify the exact gap or\nerror, and plan the minimal change. Do not alter unrelated layers.\n\nSummarise gathered requirements or findings.\n\n```json\n{\"contract\":{\"type\":\"comment\",\"comment\":{\"min_length\":80},\"required\":true}}\n```\n\n## Draft Markdown\n\n### For `create` — draft new adapter markdown\n\nDraft the full markdown for `train`.\n\n### For `review` — draft findings report\n\nWrite a structured gap report to `$REVIEW_FINDINGS_FILE` (an absolute path\nprovided by the invoking agent or chosen by you under\n`$KAIROS_LOCAL_ARTIFACT_DIR` with a session-unique segment). For each finding\ninclude:\n\n- adapter title\n- gap type\n- severity\n- why the issue matters for AI-agent behaviour\n- specific recommended fix\n\n### For `fix` — draft corrected adapter markdown\n\nApply the identified fix and show a diff-style summary of what changed.\n\n### Shared drafting rules (all operations)\n\n**Core KAIROS rule — right mode at the right time:**\n\nEvery protocol must help the agent recognise whether the current step is for\nexploration, clarification, ideation, planning, execution, validation, review,\nor handoff. Do not blur those modes.\n\n**Truth over fluency:**\n\nDesign protocols to reduce guessing and fake certainty. Prefer:\n- explicit prompts\n- explicit stop conditions\n- explicit uncertainty escalation\n- observable verification over narrative reassurance\n\n**User need over literal phrasing:**\n\nThe protocol should help the agent provide what the user actually needs, not\njust echo the wording used in the request.\n\n**DRY — Don't Repeat Yourself:**\n\nEvery piece of knowledge must live in exactly one place in the protocol family.\nIf two protocols share >50% of their body, extract shared logic.\n\n**350-line maximum per file:**\n\nNo protocol file may exceed 350 lines. If it grows beyond that, split it into\nlinked adapters or extract a reference file.\n\n**Generic-first — create generic, link from detailed:**\n\nCreate protocols from generic to detailed. The generic adapter must stand alone.\nExtensions should contain only the delta.\n\n**Slug convention — deterministic protocol linking:**\n\nEvery protocol gets a short, unique, lowercase-hyphenated `slug`.\n\n**Artifact workflow for adapters:**\n\nWhen protocol execution depends on scripts or config blobs, store them as\nartifacts linked to the adapter via `train` with:\n- `mime` set to a non-markdown content type\n- `artifact_name` set to a stable file name\n- `adapter_uri` set to the target adapter URI\n\nRetrieve attached artifacts with `export` using\n`kairos://artifact/{uuid}` when a step needs the raw content.\n\n**Protocol structure rules:**\n- H1: protocol title\n- First H2: Activation Patterns\n- Second H2: Preflight Dependencies\n- Middle H2s: one per step, each ending with a JSON contract block\n- Last H2: Reward Signal\n- Frontmatter: `slug` (required), `version` (optional), `chain_root` (required for mid-chain adapters — slug of the chain entry point)\n\n**Challenge types:**\n\nLoad `challenge-type-guide` for decision rules, JSON formats, and interpreter\nselection (`forward` with `kairos://adapter/challenge-type-guide`).\n\n**If Router pattern:** draft the router and each extension as separate files.\n\nPost the full drafted markdown.\n\n```json\n{\"contract\":{\"type\":\"comment\",\"comment\":{\"min_length\":120},\"required\":true}}\n```\n\n## User Review\n\nFor `create`: present the drafted protocol(s) and ask for approval, changes, or\ncancellation.\n\nFor `review`: present the findings report and ask which findings to act on.\n\nFor `fix`: present the corrected adapter markdown and a summary of changes.\n\n```json\n{\"contract\":{\"type\":\"user_input\",\"user_input\":{\"prompt\":\"Approve the draft (create/fix) or acknowledge findings (review)\"},\"required\":true}}\n```\n\n## Finalise Operation\n\nFor `create`: call `train` with the approved markdown. Use `force_update: false`\nunless overwrite was explicitly requested.\n\nFor `review`: do not call `train`. Save the findings report only if requested.\n\nFor `fix`: call `train` with `force_update: true` and update the corresponding\nrepo file to keep the local copy in sync.\n\nReport the outcome: adapter URI(s) for create / fix, or findings summary for\nreview.\n\n```json\n{\"contract\":{\"type\":\"comment\",\"comment\":{\"min_length\":80},\"required\":true}}\n```\n\n## Reward Signal\n\nProtocol complete when:\n- `create` — adapter(s) stored via `train` and URI(s) reported\n- `review` — findings report delivered and acknowledged\n- `fix` — corrected adapter trained with `force_update: true` and repo file updated\n\nA successful result means the resulting KAIROS adapter is more truthful,\nphase-aware, agent-readable, and aligned with human–AI harmony.\n",
     "refine-search": "---\nversion: \"4.5.1\"\nslug: refine-search\ntitle: Get help refining your search\n---\n\n\n\n# Get help refining your search\n\n**You are an AI agent.** You ran `activate` and got no solid match, or only weak\n/ ambiguous matches. This adapter helps you turn a vague request into a better\nquery so the next `activate` can find the right adapter.\n\nKAIROS principle: before acting, understand the user better. The goal is not to\nrepeat the failed phrase with minor edits. The goal is to recover the user's\nreal intent.\n\n## Activation Patterns\n\n**Run this adapter when:**\n- `activate` returned no match.\n- `activate` returned only weak choices.\n- The response included a `refine` choice.\n\n**Trigger pattern:** after `activate` when no strong match exists.\n\n**Must Never:**\n- Run when a strong match already exists.\n- Loop refinement endlessly.\n\n**Must Always:**\n- Infer the real goal, context, and missing constraints from the user's request.\n- Build a more specific query than the one that failed.\n- Run at most once per user request; if the second activation still fails, stop\n  and offer clarification or protocol creation.\n\n**Good trigger examples:**\n- activate returned only \"refine\" / weak matches → run this adapter\n- \"No protocol matched 'do the thing'\" → run this adapter\n\n**Bad trigger examples:**\n- activate returned a strong match → use that match instead\n- user asked \"what adapters exist?\" → use `activate` / `spaces`, not this\n\n## Step 1: Extract what the user actually wants\n\nFrom the user's original message, identify:\n- the real goal\n- relevant context\n- missing constraints\n- likely wording that would help KAIROS find the right adapter\n\nWrite your analysis as the solution.\n\n```json\n{\n  \"contract\": {\n    \"type\": \"comment\",\n    \"comment\": { \"min_length\": 60 },\n    \"required\": true\n  }\n}\n```\n\n## Step 2: Build and run a refined activate query\n\nUsing Step 1, construct one refined query with 3–8 specific words. Do not reuse\nthe vague phrase that already failed.\n\nCall `activate` with it.\n\n- **Strong match (score >= 0.5):** pick that choice and `forward`.\n- **Weak matches only:** ask the user to clarify or offer the create path.\n\nDo not loop more than once.\n\n```json\n{\n  \"contract\": {\n    \"type\": \"mcp\",\n    \"mcp\": { \"tool_name\": \"activate\" },\n    \"required\": true\n  }\n}\n```\n\n## Reward Signal\n\nOnly reachable after all prior steps are solved.\n\nThe agent has converted a failed literal search into a more truthful,\nuser-aligned activation attempt.\n",
     "create-new-protocol-review": "---\nversion: \"4.5.1\"\nslug: create-new-protocol-review\ntitle: Review and Publish New KAIROS Protocol\n---\n\n\n# Review and Publish New KAIROS Protocol\n\nValidate a drafted KAIROS protocol through format review, stranger review, user\napproval, and train. Invoked by the authoring adapter (`create-new-protocol`) or\ndirectly to re-review an existing draft.\n\n## Activation Patterns\n\n**Typically invoked by:** the authoring adapter's handoff step via\n`activate()` or a stored adapter URI.\n\n**Can be invoked directly when user says:**\n- \"review my protocol draft\" / \"validate this protocol\"\n- \"re-run format review on my draft\"\n- \"train my protocol\" / \"register my adapter\"\n\n**Trigger pattern:** **review** / **validate** / **train** + (protocol / draft).\n\n**Must Never:**\n- Run shell verification without verifying that all input file paths exist first.\n- Skip format or stranger review.\n- Approve a draft merely because it sounds polished.\n\n**Must Always:**\n- Verify all input file paths exist before proceeding (`test -f`).\n- Require both reviews to pass before presenting to the user.\n- Freeze the draft checksum before `train`.\n- Review for truth, phase clarity, ambiguity, and weaker-agent executability.\n\n**Good trigger examples:**\n- handoff from `create-new-protocol` → run this protocol\n- \"Review and train my protocol draft\" → run this protocol\n\n**Bad trigger examples:**\n- \"Create a new protocol\" → use `create-new-protocol`\n- \"Search for a protocol\" → use `activate`\n\n## Preflight Dependencies\n\nBefore running review/publish steps, verify prerequisites up front:\n\n1. Required MCP tools are reachable: `activate`, `forward`, `reward`, `train`.\n2. Required shell utilities used in this workflow are callable.\n3. Required authentication/session state is valid for MCP and SCM hosts.\n4. All input file paths provided by the parent agent exist (`test -f`).\n\nIf any dependency or auth check fails, stop and ask the user to choose either:\nproceed with AI-assisted setup or abort this run now.\n\n```json\n{\"contract\":{\"type\":\"comment\",\"comment\":{\"min_length\":80},\"required\":true}}\n```\n\n## Prerequisites\n\nThis protocol receives all file paths as absolute paths from the invoking agent:\n\n- `$DRAFT_FILE` — the protocol draft to review (required)\n- `$REQUIREMENTS_FILE` — requirements document (optional)\n- `$FORMAT_REVIEW_FILE` — output path for format review verdict\n- `$STRANGER_REVIEW_FILE` — output path for stranger review verdict\n- `$DRAFT_CHECKSUM_FILE` — output path for the frozen checksum\n\nThe parent agent is responsible for choosing collision-free paths (e.g., under\n`$KAIROS_LOCAL_ARTIFACT_DIR` with a session-unique segment). This protocol only\nverifies paths exist via `test -f` and writes to the provided output paths.\n\nIf invoked directly, export the variables before proceeding.\n\n**Security:** never write credentials, tokens, or secrets into output files.\n\n## Format Review [SUBAGENT]\n\nDelegate format verification to a subagent. The subagent writes its verdict to\n`$FORMAT_REVIEW_FILE`.\n\n**Input:** `$DRAFT_FILE`\n\n**Subagent task:**\n1. Read the draft.\n2. Verify structure:\n   - H1 present exactly once\n   - first H2 is `Activation Patterns`\n   - last H2 is `Reward Signal`\n   - every middle H2 has a challenge block\n   - no duplicate H2 headings\n3. Verify authoring quality:\n   - explicit phase separation where relevant\n   - no hidden assumptions\n   - no vague filler such as \"as needed\" when a concrete instruction is required\n4. Verify challenge type selection using `challenge-type-guide`:\n   - no `comment` challenges for steps that create files, call tools, or check state\n   - `shell` challenges include `timeout_seconds`\n   - shell chains use `&&`, not `;`\n5. Verify DRY, slug, and 350-line compliance.\n6. Write verdict to `$FORMAT_REVIEW_FILE`. First line MUST be\n   `PASS` or `FAIL`. Remaining lines list concrete fixes.\n\n**Output:** `$FORMAT_REVIEW_FILE`\n\n```json\n{\"contract\":{\"type\":\"shell\",\"shell\":{\"cmd\":\"test -f \\\"$FORMAT_REVIEW_FILE\\\" && head -1 \\\"$FORMAT_REVIEW_FILE\\\" | grep -qi 'pass'\",\"timeout_seconds\":5},\"required\":true}}\n```\n\n## Stranger Review [SUBAGENT]\n\nDelegate executability review to a subagent with junior AI agent persona. The\nsubagent writes its verdict to `$STRANGER_REVIEW_FILE`.\n\n**Input:** `$DRAFT_FILE`\n\n**Subagent task:**\n1. Assume persona: junior AI agent, first time encountering KAIROS protocols.\n2. Mocked dry-run: walk every step in order and produce structurally correct\n   challenge responses.\n3. Check:\n   - can I execute this without guessing?\n   - are phases clear?\n   - do I know when to stop, ask, execute, verify, or escalate uncertainty?\n   - are any variables or concepts undefined?\n4. Write verdict to `$STRANGER_REVIEW_FILE`. First line MUST be\n   `PASS` or `FAIL`. If FAIL, include specific failing mocks and missing context.\n\n**Output:** `$STRANGER_REVIEW_FILE`\n\n```json\n{\"contract\":{\"type\":\"shell\",\"shell\":{\"cmd\":\"test -f \\\"$STRANGER_REVIEW_FILE\\\" && head -1 \\\"$STRANGER_REVIEW_FILE\\\" | grep -qi 'pass'\",\"timeout_seconds\":5},\"required\":true}}\n```\n\n## Checksum Freeze [SUBAGENT]\n\nFreeze the draft after both reviews pass.\n\n**Input:** `$DRAFT_FILE`\n\n**Actions:**\n1. Compute md5 checksum of the draft file.\n2. Write checksum to `$DRAFT_CHECKSUM_FILE`.\n\n**Output:** `$DRAFT_CHECKSUM_FILE`\n\n```json\n{\"contract\":{\"type\":\"shell\",\"shell\":{\"cmd\":\"md5 -q \\\"$DRAFT_FILE\\\" > \\\"$DRAFT_CHECKSUM_FILE\\\" && test -s \\\"$DRAFT_CHECKSUM_FILE\\\"\",\"timeout_seconds\":5},\"required\":true}}\n```\n\n## User Review [SUBAGENT]\n\nPresent the reviewed protocol to the user for final approval.\n\n**Input:** draft plus both review files\n\n**Actions:**\n1. Present full protocol markdown.\n2. Summarise pattern used, step count, challenge types, slug(s), and both review verdicts.\n3. Ask the user to approve, request changes, or cancel.\n\n**Output:** user approval, changes, or cancellation.\n\n```json\n{\"contract\":{\"type\":\"user_input\",\"user_input\":{\"prompt\":\"Review the protocol draft above. Type 'approved' to run train, describe adjustments, or 'cancel'.\"},\"required\":true}}\n```\n\n## Train adapter [SUBAGENT]\n\nVerify the draft is unchanged since checksum freeze, then call `train`.\n\n**Input:** approved `$DRAFT_FILE`\n\n**Actions:**\n1. Verify draft unchanged against checksum.\n2. Call `train` with `force_update: false` unless overwrite was explicitly requested.\n3. If router pattern: train router first, then extensions in order.\n4. Verify each `train` call returned a URI.\n5. Report stored adapter URI(s) back to the user.\n\n**Output:** confirmation with stored adapter URI(s)\n\n```json\n{\"contract\":{\"type\":\"mcp\",\"mcp\":{\"tool_name\":\"train\"},\"required\":true}}\n```\n\n## Reward Signal\n\nOnly reachable after all prior steps are solved.\n\nThe user now has a trained KAIROS adapter that passed:\n1. format review\n2. stranger review\n3. checksum verification\n4. user approval\n5. successful storage via `train`\n\nThe resulting adapter is more likely to be truthful, phase-aware, and usable by\nweaker agents.\n",
-    "challenge-type-guide": "---\nversion: \"4.5.1\"\nslug: challenge-type-guide\ntitle: Challenge Type Selection Guide\n---\n\n\n\n# Challenge Type Selection Guide\n\nDecision rules, JSON formats, interpreter selection, and anti-patterns for\nchoosing KAIROS challenge types. Loaded by protocol-authoring agents during\ndrafting and review.\n\n## Activation Patterns\n\n**Typically invoked by:** `create-new-protocol` and\n`create-new-protocol-review`.\n\n**Can be invoked directly when agent needs:**\n- \"challenge type reference\" / \"which challenge type to use\"\n- \"how to write shell challenges\" / \"challenge format\"\n- \"interpreter selection for challenges\"\n\n**Trigger pattern:** **challenge** + (type | format | selection | guide).\n\n**Must Never:**\n- Be used as an execution protocol.\n- Select `comment` merely because it is easy.\n\n**Must Always:**\n- Be consulted before assigning challenge types to protocol steps.\n- Choose the strongest observable contract available.\n- Match challenge choice to the mode of the step: ask, decide, act, verify, or review.\n\n**Good trigger examples:**\n- drafting a new protocol and need challenge types → load this\n- reviewing challenge type selection in a draft → load this\n\n**Bad trigger examples:**\n- \"Create a new protocol\" → use `create-new-protocol`\n- \"Review my protocol draft\" → use `create-new-protocol-review`\n\n## Core Rules\n\n`comment` is the challenge type of **last resort**. If a step produces any\nobservable artifact — a file, a git state change, an API response, or an exit\ncode — use `shell` or `mcp` instead.\n\n**Mode before challenge type:**\n- **clarify / approve** → `user_input`\n- **call a KAIROS / MCP tool** → `mcp`\n- **execute or verify local state** → `shell`\n- **pure reasoning with no observable artifact** → `comment`\n\n### Decision Tree\n\n```\nDoes the step call an MCP tool?\n ├─ YES → Is the step a chain link to another adapter?\n │         ├─ YES → \"mcp\" with tool_name + arguments (forward + slug URI)\n │         └─ NO  → \"mcp\" with tool_name only\n └─ NO → Does the step produce or verify a file, command result, or system state?\n          ├─ YES → \"shell\"\n          └─ NO → Does the step require human approval or choice?\n                   ├─ YES → \"user_input\"\n                   └─ NO → \"comment\" only if there is no stronger observable contract\n```\n\n### Challenge JSON Formats\n\n**shell** (v2 unified envelope format)\n```json\n{\"contract\":{\"type\":\"shell\",\"shell\":{\"cmd\":\"<command>\",\"timeout_seconds\":30},\"required\":true}}\n```\n\n**mcp** (v2 unified envelope format)\n```json\n{\"contract\":{\"type\":\"mcp\",\"mcp\":{\"tool_name\":\"<tool>\"},\"required\":true}}\n```\n\n**mcp with arguments (chain link)** (v2 unified envelope format)\n```json\n{\"contract\":{\"type\":\"mcp\",\"mcp\":{\"tool_name\":\"forward\",\"arguments\":{\"uri\":\"kairos://adapter/<slug>\"}},\"required\":true}}\n```\n\nUse the `arguments` form when the step's purpose is to invoke a specific\nadapter via `forward`. The server validates that `solution.mcp.arguments`\nis a superset of the contract's `mcp.arguments` (subset/deep matching),\nproving the agent called the correct target — not just any `forward` call.\n\n**user_input** (v2 unified envelope format)\n```json\n{\"contract\":{\"type\":\"user_input\",\"user_input\":{\"prompt\":\"<specific question>\"},\"required\":true}}\n```\n\n**comment** (v2 unified envelope format)\n```json\n{\"contract\":{\"type\":\"comment\",\"comment\":{\"min_length\":50},\"required\":true}}\n```\n\n### Solution JSON Formats (v2 unified envelope)\n\nEvery solution, regardless of type, follows the same envelope:\n\n```json\n{\n  \"type\": \"<challenge_type>\",\n  \"outcome\": \"success | failure | skipped\",\n  \"evidence\": { /* type-specific proof data */ },\n  \"nonce\": \"<echo from contract>\",\n  \"proof_hash\": \"<echo from contract>\"\n}\n```\n\n**shell solution**\n```json\n{\n  \"type\": \"shell\",\n  \"outcome\": \"success\",\n  \"evidence\": {\n    \"exit_code\": 0,\n    \"stdout\": \"...\",\n    \"stderr\": \"\",\n    \"duration_seconds\": 1.2\n  }\n}\n```\n\n**mcp solution**\n```json\n{\n  \"type\": \"mcp\",\n  \"outcome\": \"success\",\n  \"evidence\": {\n    \"tool_name\": \"activate\",\n    \"arguments\": { \"query\": \"example search\" },\n    \"response\": { \"choices\": [...] }\n  }\n}\n```\n\n**user_input solution**\n```json\n{\n  \"type\": \"user_input\",\n  \"outcome\": \"success\",\n  \"evidence\": {\n    \"confirmation\": \"agreed. proceed\",\n    \"timestamp\": \"2026-05-18T14:49:00Z\"\n  }\n}\n```\n\n**comment solution**\n```json\n{\n  \"type\": \"comment\",\n  \"outcome\": \"success\",\n  \"evidence\": {\n    \"text\": \"Verified all 78 issues now have repo: labels.\"\n  }\n}\n```\n\n## Interpreter Selection\n\n| Task | Use bash | Use perl | Use python3 |\n|---|---|---|---|\n| File exists, tool installed, git state | yes | no | no |\n| Simple content grep | yes | no | no |\n| Multi-line regex | no | yes | no |\n| Paragraph processing | no | yes | no |\n| JSON validation | maybe | yes | yes |\n| YAML / complex data structure checks | no | no | yes |\n\n**Rule of thumb:** if bash requires `while read`, `grep | awk`, or heavy quote\nescaping, rewrite in Perl. If you need imports for data formats, use Python.\n\n```json\n{\"contract\":{\"type\":\"comment\",\"comment\":{\"min_length\":30},\"required\":true}}\n```\n\n## Security and Working Directory Rules\n\nUse `$KAIROS_LOCAL_ARTIFACT_DIR` as the base directory for local handoff\nartifacts. File paths within it are provided as variables by the invoking agent\nto avoid collisions between parallel sessions.\n\n**Rules:**\n- all file paths in shell challenges reference variables (e.g., `$DRAFT_FILE`,\n  `$VERDICT_FILE`) set by the parent agent — never hardcode filenames\n- the parent agent constructs collision-free paths under\n  `$KAIROS_LOCAL_ARTIFACT_DIR` (e.g., with a session ID or random suffix)\n- when using KAIROS MCP, the latest `activate` / `forward` / `next` response field `kairos_local_artifact_dir` is an ordered array of URI hints (`project://<rel>`, `user://<rel>`); pick one (`project://` when you have exactly one project context, otherwise `user://`), resolve on your machine, and `export KAIROS_LOCAL_ARTIFACT_DIR=\"<absolute>\"` if your shell does not already define it\n- if the server-returned path is not usable on your local filesystem, resolve\n  locally to `$PROJECT_DIR/.local/kairos/work` and keep it stable for the run\n- prerequisites create the directory\n- Reward Signal cleans it up\n- never write credentials or secrets to files\n\n```json\n{\"contract\":{\"type\":\"comment\",\"comment\":{\"min_length\":30},\"required\":true}}\n```\n\n## Composition Patterns\n\n**Multi-check compound command**\n- chain with `&&`\n- let the first failure stop execution\n\n**Freeze-then-verify**\n- step N writes checksum\n- step N+1 verifies checksum before publish\n\n**Write-then-grep**\n- subagent writes a verdict file\n- main agent checks the first line for `PASS`\n\n**Phase-batch with critic**\n- batch related TODO items into one phase\n- verify hard artifacts with shell\n- then call `phase-critic` for adversarial review\n\n**Dry-run before execution**\n- dry-run command\n- human review\n- real command\n\n**Chain-link via forward + slug**\n- layer N's contract requires `forward` with a specific adapter slug\n- agent calls `forward` to start the next adapter (new `execution_id`)\n- next adapter runs its own layer chain through to `reward`\n- use when the protocol exceeds 350 lines or needs multi-path routing\n- NOT a replacement for layers — layers remain the primary unit of work\n- example contract:\n  ```json\n  {\"contract\":{\"type\":\"mcp\",\"mcp\":{\"tool_name\":\"forward\",\"arguments\":{\"uri\":\"kairos://adapter/implement-plan\"}},\"required\":true}}\n  ```\n\n```json\n{\"contract\":{\"type\":\"comment\",\"comment\":{\"min_length\":30},\"required\":true}}\n```\n\n## Reward Signal\n\nOnly reachable after all prior steps are solved.\n\nThe agent can now:\n1. pick the correct challenge type for each protocol step\n2. match challenge type to the step's mode\n3. choose the right interpreter\n4. apply working-directory and security rules\n5. use stronger verification patterns instead of weak narrative checks\n",
+    "challenge-type-guide": "---\nversion: \"4.5.1\"\nslug: challenge-type-guide\ntitle: Challenge Type Selection Guide\n---\n\n\n\n# Challenge Type Selection Guide\n\nDecision rules, JSON formats, interpreter selection, and anti-patterns for\nchoosing KAIROS challenge types. Loaded by protocol-authoring agents during\ndrafting and review.\n\n## Activation Patterns\n\n**Typically invoked by:** `create-new-protocol` and\n`create-new-protocol-review`.\n\n**Can be invoked directly when agent needs:**\n- \"challenge type reference\" / \"which challenge type to use\"\n- \"how to write shell challenges\" / \"challenge format\"\n- \"interpreter selection for challenges\"\n\n**Trigger pattern:** **challenge** + (type | format | selection | guide).\n\n**Must Never:**\n- Be used as an execution protocol.\n- Select `comment` merely because it is easy.\n\n**Must Always:**\n- Be consulted before assigning challenge types to protocol steps.\n- Choose the strongest observable contract available.\n- Match challenge choice to the mode of the step: ask, decide, act, verify, or review.\n\n**Good trigger examples:**\n- drafting a new protocol and need challenge types → load this\n- reviewing challenge type selection in a draft → load this\n\n**Bad trigger examples:**\n- \"Create a new protocol\" → use `create-new-protocol`\n- \"Review my protocol draft\" → use `create-new-protocol-review`\n\n## Core Rules\n\n`comment` is the challenge type of **last resort**. If a step produces any\nobservable artifact — a file, a git state change, an API response, or an exit\ncode — use `shell` or `mcp` instead.\n\n**Mode before challenge type:**\n- **clarify / approve** → `user_input`\n- **call a KAIROS / MCP tool** → `mcp`\n- **execute or verify local state** → `shell`\n- **pure reasoning with no observable artifact** → `comment`\n\n### Decision Tree\n\n```\nDoes the step call an MCP tool?\n ├─ YES → Is the step a chain link to another adapter?\n │         ├─ YES → \"mcp\" with tool_name + arguments (forward + slug URI)\n │         └─ NO  → \"mcp\" with tool_name only\n └─ NO → Does the step produce or verify a file, command result, or system state?\n          ├─ YES → \"shell\"\n          └─ NO → Does the step require human approval or choice?\n                   ├─ YES → \"user_input\"\n                   └─ NO → \"comment\" only if there is no stronger observable contract\n```\n\n### Challenge JSON Formats\n\n**shell** (v2 unified envelope format)\n```json\n{\"contract\":{\"type\":\"shell\",\"shell\":{\"cmd\":\"<command>\",\"timeout_seconds\":30},\"required\":true}}\n```\n\n**mcp** (v2 unified envelope format)\n```json\n{\"contract\":{\"type\":\"mcp\",\"mcp\":{\"tool_name\":\"<tool>\"},\"required\":true}}\n```\n\n**mcp with arguments (chain link)** (v2 unified envelope format)\n```json\n{\"contract\":{\"type\":\"mcp\",\"mcp\":{\"tool_name\":\"forward\",\"arguments\":{\"uri\":\"kairos://adapter/<slug>\"}},\"required\":true}}\n```\n\nUse the `arguments` form when the step's purpose is to invoke a specific\nadapter via `forward`. The server validates that `solution.mcp.arguments`\nis a superset of the contract's `mcp.arguments` (subset/deep matching),\nproving the agent called the correct target — not just any `forward` call.\n\n**user_input** (v2 unified envelope format)\n```json\n{\"contract\":{\"type\":\"user_input\",\"user_input\":{\"prompt\":\"<specific question>\"},\"required\":true}}\n```\n\n**comment** (v2 unified envelope format)\n```json\n{\"contract\":{\"type\":\"comment\",\"comment\":{\"min_length\":50},\"required\":true}}\n```\n\n### Solution JSON Formats (v2 unified envelope)\n\nEvery solution, regardless of type, follows the same envelope:\n\n```json\n{\n  \"type\": \"<challenge_type>\",\n  \"outcome\": \"success | failure | skipped\",\n  \"evidence\": { /* type-specific proof data */ },\n  \"nonce\": \"<echo from contract>\",\n  \"proof_hash\": \"<echo from contract>\"\n}\n```\n\n**shell solution**\n```json\n{\n  \"type\": \"shell\",\n  \"outcome\": \"success\",\n  \"evidence\": {\n    \"exit_code\": 0,\n    \"stdout\": \"...\",\n    \"stderr\": \"\",\n    \"duration_seconds\": 1.2\n  }\n}\n```\n\n**mcp solution**\n```json\n{\n  \"type\": \"mcp\",\n  \"outcome\": \"success\",\n  \"evidence\": {\n    \"tool_name\": \"activate\",\n    \"arguments\": { \"query\": \"example search\" },\n    \"response\": { \"choices\": [...] }\n  }\n}\n```\n\n**user_input solution**\n```json\n{\n  \"type\": \"user_input\",\n  \"outcome\": \"success\",\n  \"evidence\": {\n    \"confirmation\": \"agreed. proceed\",\n    \"timestamp\": \"2026-05-18T14:49:00Z\"\n  }\n}\n```\n\n**comment solution**\n```json\n{\n  \"type\": \"comment\",\n  \"outcome\": \"success\",\n  \"evidence\": {\n    \"text\": \"Verified all 78 issues now have repo: labels.\"\n  }\n}\n```\n\n### Format migration notes\n\n**Always use v2 format** when authoring new protocols or submitting solutions. The server accepts both formats:\n\n- **v2 (preferred)**: `{type, outcome, evidence: {...}}` - unified envelope for all types\n- **v1 (older)**: `{type, [type]: {...}}` - still accepted but never advertised\n\nKey v2 improvements:\n- Consistent structure across all challenge types\n- `outcome` at top level instead of nested `success` boolean\n- `evidence` envelope replaces type-specific objects\n- MCP uses `response` instead of `result`\n\nExisting protocols using v1 format continue to work unchanged. New protocols should follow v2 examples above.\n\n## Interpreter Selection\n\n| Task | Use bash | Use perl | Use python3 |\n|---|---|---|---|\n| File exists, tool installed, git state | yes | no | no |\n| Simple content grep | yes | no | no |\n| Multi-line regex | no | yes | no |\n| Paragraph processing | no | yes | no |\n| JSON validation | maybe | yes | yes |\n| YAML / complex data structure checks | no | no | yes |\n\n**Rule of thumb:** if bash requires `while read`, `grep | awk`, or heavy quote\nescaping, rewrite in Perl. If you need imports for data formats, use Python.\n\n```json\n{\"contract\":{\"type\":\"comment\",\"comment\":{\"min_length\":30},\"required\":true}}\n```\n\n## Security and Working Directory Rules\n\nUse `$KAIROS_LOCAL_ARTIFACT_DIR` as the base directory for local handoff\nartifacts. File paths within it are provided as variables by the invoking agent\nto avoid collisions between parallel sessions.\n\n**Rules:**\n- all file paths in shell challenges reference variables (e.g., `$DRAFT_FILE`,\n  `$VERDICT_FILE`) set by the parent agent — never hardcode filenames\n- the parent agent constructs collision-free paths under\n  `$KAIROS_LOCAL_ARTIFACT_DIR` (e.g., with a session ID or random suffix)\n- when using KAIROS MCP, the latest `activate` / `forward` / `next` response field `kairos_local_artifact_dir` is an ordered array of URI hints (`project://<rel>`, `user://<rel>`); pick one (`project://` when you have exactly one project context, otherwise `user://`), resolve on your machine, and `export KAIROS_LOCAL_ARTIFACT_DIR=\"<absolute>\"` if your shell does not already define it\n- if the server-returned path is not usable on your local filesystem, resolve\n  locally to `$PROJECT_DIR/.local/kairos/work` and keep it stable for the run\n- prerequisites create the directory\n- Reward Signal cleans it up\n- never write credentials or secrets to files\n\n```json\n{\"contract\":{\"type\":\"comment\",\"comment\":{\"min_length\":30},\"required\":true}}\n```\n\n## Composition Patterns\n\n**Multi-check compound command**\n- chain with `&&`\n- let the first failure stop execution\n\n**Freeze-then-verify**\n- step N writes checksum\n- step N+1 verifies checksum before publish\n\n**Write-then-grep**\n- subagent writes a verdict file\n- main agent checks the first line for `PASS`\n\n**Phase-batch with critic**\n- batch related TODO items into one phase\n- verify hard artifacts with shell\n- then call `phase-critic` for adversarial review\n\n**Dry-run before execution**\n- dry-run command\n- human review\n- real command\n\n**Chain-link via forward + slug**\n- layer N's contract requires `forward` with a specific adapter slug\n- agent calls `forward` to start the next adapter (new `execution_id`)\n- next adapter runs its own layer chain through to `reward`\n- use when the protocol exceeds 350 lines or needs multi-path routing\n- NOT a replacement for layers — layers remain the primary unit of work\n- example contract:\n  ```json\n  {\"contract\":{\"type\":\"mcp\",\"mcp\":{\"tool_name\":\"forward\",\"arguments\":{\"uri\":\"kairos://adapter/implement-plan\"}},\"required\":true}}\n  ```\n\n```json\n{\"contract\":{\"type\":\"comment\",\"comment\":{\"min_length\":30},\"required\":true}}\n```\n\n## Reward Signal\n\nOnly reachable after all prior steps are solved.\n\nThe agent can now:\n1. pick the correct challenge type for each protocol step\n2. match challenge type to the step's mode\n3. choose the right interpreter\n4. apply working-directory and security rules\n5. use stronger verification patterns instead of weak narrative checks\n",
     "phase-critic": "---\nversion: \"4.5.1\"\nslug: phase-critic\ntitle: Phase Critic\n---\n\n\n\n# Phase Critic\n\nBounded adversarial investigator fired at phase boundaries. Receives a reference\ndocument, loads the calling protocol's invariants, and independently verifies\nthe verification target by extracting claims, gathering evidence, and searching\nfor contradictions.\n\n## Activation Patterns\n\n**Typically invoked by:** another protocol at the end of a plan, implement,\nvalidation, or review phase via `forward` with the stored adapter URI\n`kairos://adapter/phase-critic`.\n\n**Can be invoked directly when agent needs:**\n- \"review phase output\" / \"verify plan\" / \"audit implementation\"\n- \"adversarial review\" / \"critic review\"\n\n**Trigger pattern:** **review** / **verify** / **audit** / **critic** +\n(phase / plan / implementation / output).\n\n**Must Never:**\n- Modify any file outside the provided input and output paths.\n- Trust the executor's assertions without independent evidence.\n- Classify a claim as `verified` based only on absence of contradiction.\n- Keep investigating after a precondition failure that blocks verification.\n\n**Must Always:**\n- Extract claims before investigating.\n- Verify in-scope claims first.\n- Search for contradictory evidence, not only confirming evidence.\n- Include confidence in every verdict and failure class in every FAIL verdict.\n- Preserve the boundary between execution and review.\n\n**Good trigger examples:**\n- implement protocol fires this after planning → run this protocol\n- compose protocol fires this after output generation → run this protocol\n\n**Bad trigger examples:**\n- \"Create a new protocol\" → use `create-new-protocol`\n- \"Implement a feature\" → use an execution protocol, not this one\n\n## Artifact Contract\n\nThis protocol receives:\n\n```text\n{\n  input_file: string,              // absolute path to the document under review\n  calling_protocol_slug: string,\n  verification_target: string,\n  context: {\n    operation: string,\n    verdict_file: string           // absolute path where the verdict will be written\n  }\n}\n```\n\nThis protocol produces:\n\n```text\n{\n  verdict: \"PASS\" | \"FAIL\",\n  confidence: \"High\" | \"Medium\" | \"Low\",\n  failure_class: string,\n  verdict_file: \"{context.verdict_file}\"\n}\n```\n\n## Precondition Check [SUBAGENT]\n\nVerify foundational preconditions before deep investigation. Early-abort if any\nare broken.\n\n**Input:** `$INPUT_FILE` (absolute path), calling protocol slug, context\n\n**Actions:**\n1. Verify the input file exists: `test -f \"$INPUT_FILE\"`.\n2. Verify the verdict output path is writable (parent directory exists).\n3. Load the calling protocol's Must Never / Must Always sections if possible.\n4. Verify the target state is available for review.\n5. Verify required tools are available.\n\n**On early-abort:**\n- verdict = `FAIL`\n- failure class = `verification_blocked`\n- confidence = `Low`\n- write focused diagnostics\n- skip all later steps\n\n**Output:** verified preconditions or early-abort diagnostics\n\n```json\n{\"contract\":{\"type\":\"shell\",\"shell\":{\"cmd\":\"test -f \\\"$INPUT_FILE\\\"\",\"timeout_seconds\":3},\"required\":true}}\n```\n\n## Audit [SUBAGENT]\n\nPerform bounded adversarial review of the target against the reference document\nand invariants.\n\n**Input:** reference document, invariants, verification target, context\n\n**Actions:**\n\n**Step 1 — Extract claims**\nExtract every verifiable claim that matters operationally:\n- workflows, commands, tool calls\n- configuration and environment variable claims\n- architecture claims that imply behaviour\n- guarantees and constraints\n- security, auth, and permission claims\n\n**Step 2 — Investigate with scope discipline**\n1. verify all in-scope claims first\n2. review high-risk claims with maximum depth\n3. expand beyond scope only when contradiction signals, dependencies, or\n   high-risk impact require it\n4. continue reviewing all in-scope claims even after a normal FAIL condition is met;\n   only a precondition failure stops the audit early\n\n**High-risk areas**\n- setup steps and prerequisites\n- CLI commands and flags\n- configuration keys and environment variables\n- authentication, identity, and permissions\n- deployment and release behaviour\n- external publishing actions\n- destructive or irreversible operations\n\n**Step 3 — Gather evidence using the hierarchy**\n\n| Priority | Evidence source |\n|---|---|\n| 1 | source code, config files, schemas, route definitions |\n| 2 | direct runtime verification |\n| 3 | tests |\n| 4 | MCP-queried system state |\n| 5 | external docs, web pages, linked references |\n| 6 | conversational assumption or indirect inference |\n\n**Rules:**\n- higher-priority evidence overrides lower-priority evidence\n- priority 5–6 alone is not enough for `verified`\n\n**Step 4 — Produce claim verdicts**\n\nFor each claim, record:\n- CLAIM\n- EVIDENCE\n- EVIDENCE PRIORITY\n- VERDICT: `verified` / `partially_verified` / `unverifiable` / `incorrect`\n\nA claim must not be marked `verified` without positive supporting evidence.\n\n**Step 5 — Review the phase boundary**\nCheck whether the calling protocol preserved the right mode boundary:\n- was execution reviewed independently?\n- were review and execution kept separate?\n- did the protocol switch to verification when it should have?\n- is escalation needed because certainty was overstated?\n\n**Output:** structured audit findings for the verdict step\n\n```json\n{\"contract\":{\"type\":\"comment\",\"comment\":{\"min_length\":120},\"required\":true}}\n```\n\n## Verdict [SUBAGENT]\n\nTurn the audit into a final PASS / FAIL result.\n\n**Input:** audit findings and verdict file path from context (use it as\n`$VERDICT_FILE`)\n\n**Actions:**\n1. Fail if any critical claim is incorrect.\n2. Fail if verification was blocked.\n3. Fail if the evidence base is too weak for high-risk claims.\n4. Otherwise, pass with the appropriate confidence.\n5. Write `$VERDICT_FILE` so line 1 is `PASS` or `FAIL`,\n   line 2 is `High`, `Medium`, or `Low`, and the remaining lines summarise key\n   findings.\n\n**Output:** final verdict file\n\n```json\n{\"contract\":{\"type\":\"shell\",\"shell\":{\"cmd\":\"test -f \\\"$VERDICT_FILE\\\" && head -1 \\\"$VERDICT_FILE\\\" | grep -qiE '^(PASS|FAIL)$' && sed -n '2p' \\\"$VERDICT_FILE\\\" | grep -qiE '^(High|Medium|Low)$'\",\"timeout_seconds\":5},\"required\":true}}\n```\n\n## Reward Signal\n\nOnly reachable after all prior steps are solved.\n\nThe phase output has been independently challenged. The resulting verdict is more\ntruthful because it is based on extracted claims, explicit evidence, contradiction\nsearch, and a clean boundary between execution and review.\n",
     "protocol-linking-guide": "---\nversion: \"4.5.1\"\nslug: protocol-linking-guide\ntitle: Protocol Linking Guide\n---\n\n\n\n# Protocol Linking Guide\n\nDecision rules for structuring adapter relationships: layers vs chains,\n`activate` vs `forward` + slug, chain-root collapsing, and MCP argument\nvalidation for deterministic cross-protocol linking.\n\n## Activation Patterns\n\n**Typically invoked by:** `create-new-protocol` and\n`create-new-protocol-review`.\n\n**Can be invoked directly when agent needs:**\n- \"layers vs chains\" / \"when to split into multiple adapters\"\n- \"how to link adapters\" / \"cross-protocol linking\"\n- \"activate vs forward\" / \"when to use forward with slug\"\n- \"chain-root collapsing\" / \"how chain_root works\"\n\n**Trigger pattern:** **layers** / **chains** / **linking** / **forward** +\n(slug | adapter | protocol).\n\n**Must Never:**\n- Be used as an execution protocol.\n- Recommend chaining when layers suffice.\n\n**Must Always:**\n- Be consulted before splitting a protocol into multiple adapters.\n- Be consulted before adding `forward` calls or `chain_root` frontmatter.\n\n**Good trigger examples:**\n- \"my protocol is 400 lines, should I split?\" → load this\n- \"how do I link adapter A to adapter B?\" → load this\n- \"what is chain_root?\" → load this\n\n**Bad trigger examples:**\n- \"which challenge type for this step?\" → use `challenge-type-guide`\n- \"create a new protocol\" → use `create-new-protocol`\n\n## Layers vs Chains\n\n**Layers** (H2 steps within one adapter) are the preferred unit of work. Use\nlayers when all steps fit in one adapter under 350 lines and serve a single\ncohesive concern.\n\n**Chain via `forward` + slug** — use only when:\n- the adapter exceeds 350 lines\n- weaker models cannot follow the full adapter in one pass\n- runtime routing to different adapters is needed (Router pattern)\n- a step is reusable across multiple parent adapters (e.g. `phase-critic`)\n\nChaining is NOT a replacement for layers. A 6-step protocol that fits in 350\nlines should stay as one adapter with 6 H2 layers.\n\n```json\n{\"contract\":{\"type\":\"comment\",\"comment\":{\"min_length\":30},\"required\":true}}\n```\n\n## `activate` vs `forward` + slug\n\n| Situation | Use |\n|---|---|\n| Initial user-intent resolution | `activate` |\n| Target adapter is known at authoring time | `forward` + slug |\n| Target depends on runtime classification the current adapter cannot resolve | `activate` |\n| Linking to the next adapter in a chain | `forward` + slug |\n| Reusable sub-protocol invoked by multiple parents | `forward` + slug |\n\nNever call `activate` just to get a slug you already know — it wastes a\nround-trip and introduces non-determinism.\n\n```json\n{\"contract\":{\"type\":\"comment\",\"comment\":{\"min_length\":30},\"required\":true}}\n```\n\n## Chaining Back and Forward\n\nWhen a family of adapters has a root conductor and smaller domain extensions,\ndocument both directions of chaining.\n\n### Chaining back with `chain_root`\n\nExtension adapters should include `chain_root` frontmatter and a prerequisite\nclause:\n\n**Prerequisite:** Start from `<root-adapter-slug>` unless the caller already\nprovided a mission brief, safety gates, authorization boundary, and evidence\nrequirements.\n\nThis protects activation entry points. If a user or weak agent activates a\nmid-chain extension directly, the extension can collapse back to the root unless\nit already received the required parent context.\n\n### Chaining forward with `forward`\n\n`chain_root` does not replace explicit `forward` calls inside the chain.\n\nWhen the next adapter is known at authoring time or after deterministic runtime\nclassification, use `forward` with the target slug. Do not call `activate` just\nto rediscover a slug already known by the current adapter.\n\nForwarded adapters should receive:\n- current mission brief;\n- safety gates;\n- authorization boundary;\n- evidence requirements;\n- user constraints;\n- reason for forwarding.\n\nUse `activate` only when the next adapter must be discovered from an ambiguous\nruntime need.\n\n### Multi-layer conductors and extension URIs\n\nWhen you are executing **`forward`** inside a **multi-layer** adapter (for example\na root conductor), each continuation follows **`next_call`** for the **same**\n`execution_id` until the last layer directs **`reward`**. A layer that mentions\nanother adapter URI only as routing guidance — for example\n`kairos://adapter/e2e-coding-pr-delivery` — does **not** cause the server to\nswitch executions: the next **`forward`** in that run still advances the\n**current** adapter’s next layer.\n\nTo run an extension adapter, call **`forward`** with\n`kairos://adapter/<slug>` as a **new** first call (adapter URI; omit\n`solution`). That starts a **separate** execution with its own layers. Common\npattern: complete the conductor with **`reward`**, then **`forward`** to the\nextension slug when coding or PR-shaped work should run under that adapter’s\ncontracts.\n\n```json\n{\"contract\":{\"type\":\"comment\",\"comment\":{\"min_length\":30},\"required\":true}}\n```\n\n## System adapters vs user adapters\n\nBoth system and user-authored adapters chain using slugs in `next_action`:\n\n```text\ncall forward with kairos://adapter/phase-critic ...\ncall forward with kairos://adapter/implement-terraform ...\n```\n\nSlugs are globally unique (enforced by `train`), so resolution is\ndeterministic — no scoring, no ambiguity.\n\n```json\n{\"contract\":{\"type\":\"comment\",\"comment\":{\"min_length\":30},\"required\":true}}\n```\n\n## MCP argument validation for chain links\n\nA bare `mcp` contract checks only the tool name:\n\n```json\n{\"contract\":{\"type\":\"mcp\",\"mcp\":{\"tool_name\":\"forward\"},\"required\":true}}\n```\n\nAn `mcp` contract **with `arguments`** verifies the exact target:\n\n```json\n{\"contract\":{\"type\":\"mcp\",\"mcp\":{\"tool_name\":\"forward\",\"arguments\":{\"uri\":\"kairos://adapter/code-review-policy\"}},\"required\":true}}\n```\n\nThe server validates that `solution.mcp.arguments` is a superset of the\ncontract's `mcp.arguments` (subset/deep matching). This proves the agent\ncalled `forward` with the correct slug, not just any `forward` call.\n\nUse the `arguments` form whenever a step's purpose is to invoke a specific\nknown adapter.\n\n```json\n{\"contract\":{\"type\":\"comment\",\"comment\":{\"min_length\":30},\"required\":true}}\n```\n\n## Chain-root collapsing\n\nWhen a chain has multiple adapters, `activate` might match a mid-chain adapter\ndirectly. Without protection, the agent would start at the wrong step, missing\nprerequisite context gathered by earlier phases.\n\n**`chain_root` frontmatter field** solves this at the server level. Add it to\nevery mid-chain adapter:\n\n```yaml\n---\nslug: implement-plan\nversion: 1.0.0\nchain_root: implement\n---\n```\n\n**Server behaviour:** When `activate` returns a match that has `chain_root`,\nthe server resolves the slug to the root adapter's URI and replaces the\nchoice's `uri` and `next_action` to point there. Multiple mid-chain matches\nfrom the same chain are deduplicated to a single entry. The agent always\nstarts at the chain entry point.\n\n**Defense in depth — protocol-level safeguards:** Each mid-chain adapter\nshould include a **Prerequisites** paragraph in its Activation Patterns\nlisting inputs required from earlier phases and directing the agent to start\nfrom the chain root if those inputs are missing. This catches direct `forward`\ncalls that bypass `activate`.\n\n**Rules:**\n- The chain root adapter must NOT have `chain_root` in its frontmatter.\n- `chain_root` must refer to an existing slug (the chain entry point).\n- `forward` calls within a chain are unaffected — `chain_root` only changes\n  `activate` output routing.\n\n```json\n{\"contract\":{\"type\":\"comment\",\"comment\":{\"min_length\":30},\"required\":true}}\n```\n\n## Reward Signal\n\nOnly reachable after all prior steps are solved.\n\nThe agent can now:\n1. decide when to use layers vs chains\n2. choose `activate` vs `forward` + slug correctly\n3. use slug-based linking for all adapters (system and user)\n4. apply MCP argument validation for deterministic chain links\n5. implement `chain_root` collapsing with defense-in-depth safeguards\n6. document chaining back (`chain_root` + prerequisites) and chaining forward\n   (`forward` with known slugs) for root + extension families\n",
     "bulk-insert-adapters-via-cli": "---\nslug: bulk-insert-adapters-via-cli\nversion: \"4.5.1\"\ntitle: Bulk Insert Adapters via KAIROS CLI\n---\n\n# Bulk Insert Adapters via KAIROS CLI\n\n## Activation Patterns\n\nBatch-register multiple KAIROS adapters from a local directory tree using the\n`kairos train` CLI. Handles directory scanning, validation, space targeting,\nand error recovery for bulk operations.\n\n**Run this protocol when the user says ANY of:**\n\n- \"bulk insert adapters\" / \"bulk train adapters\"\n- \"train all protocols in a directory\"\n- \"batch register adapters from folder\"\n- \"upload all .md files to KAIROS\"\n- \"sync local protocols to KAIROS\" / \"push protocols to KAIROS\"\n- \"train directory recursively\" / \"train --recursive\"\n\n**Trigger pattern:** **bulk** / **batch** / **all** / **directory** / **recursive** +\n(train | insert | register | upload | push | sync) + (adapters | protocols).\n\n**Must Never:**\n\n- Train files without first validating their structure.\n- Overwrite existing adapters without explicit user confirmation (`--force`).\n- Train files that contain secrets or token-like strings without `--allow-sensitive-upload`.\n- Skip reporting partial failures in a batch.\n\n**Must Always:**\n\n- Verify the target directory exists and contains `.md` files.\n- Show the user the file list and target space before executing.\n- Report per-file success/failure after the batch completes.\n- Respect the user's choice of space (personal or group).\n\n**Good trigger examples:**\n\n- \"Train all adapters in kairos-v4/_personal/\" → run this protocol\n- \"Bulk insert the protocols from my export folder\" → run this protocol\n- \"Push all .md files in this directory to my personal space\" → run this protocol\n\n**Bad trigger examples:**\n\n- \"Train this one protocol\" → use `kairos train <file>` directly\n- \"Create a new protocol\" → use `create-new-protocol`\n- \"Export my adapters\" → use `kairos export`\n\n## Preflight Dependencies\n\nVerify before execution:\n\n1. `kairos` CLI is installed and callable (`kairos --version`).\n2. Authentication is valid (`kairos token` returns a token without error).\n3. Target directory path exists and contains at least one `.md` file.\n4. If a space is specified, confirm it exists via `kairos spaces`.\n\nIf any check fails, stop and report the issue with remediation guidance.\n\n```json\n{\"contract\":{\"type\":\"shell\",\"shell\":{\"cmd\":\"kairos --version && kairos token > /dev/null 2>&1 && echo 'auth: ok'\",\"timeout_seconds\":15},\"required\":true}}\n```\n\n## Confirm Scope\n\nPresent the batch operation details to the user:\n\n1. **Source directory** — absolute path.\n2. **File count** — number of `.md` files found (recursive or flat).\n3. **Target space** — personal (default) or a named group space.\n4. **Force mode** — whether `--force` will be used (overwrites existing).\n5. **Model attribution** — which `--model` value to use.\n\nList the files that will be trained and ask for confirmation.\n\n```json\n{\"contract\":{\"type\":\"user_input\",\"user_input\":{\"prompt\":\"Confirm batch scope: directory, file list, target space, force mode, and model ID\"},\"required\":true}}\n```\n\n## Validate Structure\n\nBefore training, validate each `.md` file against KAIROS adapter requirements:\n\n- Has an H1 title.\n- First H2 is \"Activation Patterns\".\n- Last H2 is \"Reward Signal\".\n- Contains at least one fenced `json` block with a top-level `contract` object.\n- Does not exceed 350 lines.\n\nReport any files that fail validation. Ask user whether to:\n- skip invalid files and proceed with valid ones\n- abort and fix issues first\n\n```json\n{\"contract\":{\"type\":\"shell\",\"shell\":{\"cmd\":\"echo 'Validation complete — see output above for results'\",\"timeout_seconds\":30},\"required\":true}}\n```\n\n## Execute Bulk Train\n\nRun the batch train operation using the KAIROS CLI:\n\n```\nkairos train <directory> --recursive --model <model> --space <space> [--force]\n```\n\nIf the directory contains files that should NOT be trained (README, reference\ndocs), either:\n- train file-by-file for the valid subset, or\n- move non-adapter files out before running `--recursive`\n\nCapture stdout and stderr. The CLI reports per-file status.\n\n```json\n{\"contract\":{\"type\":\"shell\",\"shell\":{\"cmd\":\"kairos train <directory> --recursive --model <model_id> --space <space>\",\"timeout_seconds\":120},\"required\":true}}\n```\n\n## Report Results\n\nSummarise the batch outcome:\n\n| Metric | Value |\n|--------|-------|\n| Files attempted | N |\n| Successfully trained | N |\n| Failed | N |\n| Skipped (invalid) | N |\n\nFor each failure, report:\n- File path\n- Error message\n- Suggested fix\n\nIf all succeeded, confirm adapter URIs are accessible via `kairos spaces`.\n\n```json\n{\"contract\":{\"type\":\"comment\",\"comment\":{\"min_length\":80},\"required\":true}}\n```\n\n## Reward Signal\n\nProtocol complete when:\n- All valid adapter files in the target directory have been trained to the\n  specified space.\n- Per-file results have been reported to the user.\n- Any failures have been documented with remediation guidance.\n\nA successful bulk insert means the user's local protocol library is now\nregistered in KAIROS and available for activation by AI agents.\n"
